requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
newspaper3k
